requirements.txt
psycopg2-binary==2.9.9
geopandas~=0.14.0
shapely~=2.0.0
pyproj~=3.6.0
pandas~=2.2.0
numpy~=1.26.0
python-dotenv~=1.0.0
requests~=2.31.0
pytest~=8.2.0
pytest-cov~=5.0.0
click~=8.1.0
tqdm~=4.66.0
tenacity==8.2.3
osmium==3.7.0
